GPS: Remove non-working GPS-switch - contributes #7
Disabling GPS positioning from the SettingsPage inside the app
didn't really disable positioning in all situations.
SailfishOS has global GPS switch which could be used instead
to avoid confusion. Also it seems that getting the GPS location
is possible even when it's disabled globally. This might be bug
in SailfishOS.