Migrate route query to newer Digitransit API - fixes #15
From now on the same Digitransit Graphql based itinerary planning
API is used in Helsinki and Tampere. In the future its easily
possible to cover even more cities in Finland using the same API.

This is unnecessarily big commit which refactors the whole query
logic and result parsing. From now on the result is in a Singleton
and not passed around.

Digitransit API seems to be missing stop times so those are
removed from the results.